+++ lib/flutter_oembed.dart
@@ -19,6 +19,7 @@ library;
 // Configuration & Scope
 export 'src/core/embed_scope.dart';
 export 'src/core/embed_cache_provider.dart';
+export 'src/core/embed_service_interface.dart';
 export 'src/logging/embed_logger.dart';
 //
 // Models